use shared vite config
