+++ apps/demo-app/vite.config.ts
@@ -1,7 +1,7 @@
-import vue from "@vitejs/plugin-vue";
 import { defineConfig } from "vite";
+import { VITE_BASE_CONFIG } from "./node_modules/@sit-onyx/shared/vite.config.base";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
+  ...VITE_BASE_CONFIG,
 });
