+++ apps/docs/package.json
@@ -16,6 +16,7 @@
     "@playwright/test": "^1.47.2",
     "@sit-onyx/chartjs-plugin": "workspace:^",
     "@sit-onyx/icons": "workspace:^",
+    "@sit-onyx/shared": "workspace:^",
     "@sit-onyx/vitepress-theme": "workspace:^",
     "chart.js": "^4.4.4",
     "postcss": "^8.4.47",
