+++ packages/sit-onyx/package.json
@@ -50,10 +50,11 @@
   "devDependencies": {
     "@axe-core/playwright": "^4.10.0",
     "@sit-onyx/headless": "workspace:^",
+    "@sit-onyx/shared": "workspace:^",
     "@storybook/addon-a11y": "^8.3.5",
     "@vue/compiler-dom": "catalog:",
     "eslint-plugin-vue-scoped-css": "^2.8.1",
-    "sass": "catalog:",
+    "sass-embedded": "catalog:",
     "vue": "catalog:",
     "vue-component-type-helpers": "^2.1.6"
   }
