+++ packages/sit-onyx/vite.config.ts
@@ -1,15 +1,14 @@
 /// <reference types="vitest" />
 import vue from "@vitejs/plugin-vue";
 import { fileURLToPath, URL } from "node:url";
-import { Deprecation, deprecations, Options } from "sass";
 import { defineConfig } from "vite";
+import { VITE_BASE_CONFIG } from "./node_modules/@sit-onyx/shared/vite.config.base";
 import packageJson from "./package.json";
 import { vuePluginOptions } from "./playwright.config";
 
-type ViteScssOptions = Options<"async"> & { api: "modern-compiler" };
-
 // https://vitejs.dev/config
 export default defineConfig({
+  ...VITE_BASE_CONFIG,
   plugins: [vue(vuePluginOptions)],
   build: {
     lib: {
@@ -30,17 +29,6 @@ export default defineConfig({
     coverage: {
       include: ["src"],
       exclude: ["src/**/*.stories.ts"],
-    },
-  },
-  css: {
-    preprocessorOptions: {
-      scss: {
-        api: "modern-compiler",
-        // error for all warnings
-        fatalDeprecations: (Object.values(deprecations) as Deprecation[]).filter(
-          ({ status }) => status !== "future",
-        ),
-      } satisfies ViteScssOptions,
     },
   },
 });
