+++ pnpm-workspace.yaml
@@ -6,4 +6,5 @@ catalog:
   "@vue/compiler-dom": "3.5.11"
   "chart.js": "^4.4.4"
   "vue": "3.5.11"
-  "sass": "^1.79.3"
+  "sass": "1.79.4"
+  "sass-embedded": "1.79.4"
